refactor(Profiles): split ProfilesContextFacade into interface and implementation, update package structure
- Refactored `ProfilesContextFacade` into `ProfilesContextFacade` interface and `ProfilesContextFacadeImpl` implementation.
- Moved `ProfilesContextFacadeImpl` to `application.acl` package, keeping interface in `interfaces.acl` package for improved separation of concerns and flexibility in implementation.
